+++ backend/questions/models.py
@@ -4,14 +4,6 @@ from cassandra.cqlengine import columns
 from uuid import uuid4
 
 # FORMULAIRE ET REPONSES
-class ModulesESG(DjangoCassandraModel):
-    id = columns.UUID(primary_key=True, default=uuid4)
-    id_client = columns.UUID(required=True)
-    date_last_modification = columns.DateTime(required=True)  # Date of the last modification
-    original_answers = columns.List(columns.UUID(required=True), required=True)
-    modified_answers = columns.List(columns.UUID(required=True), required=False)
-    state = columns.Text(required=True)  # State of the module (e.g., 'open', 'validation', 'verified')
-    calculated_score = columns.Double(required=False)  # Calculated score of the module in validation state
 
 class CommitmentPacts(DjangoCassandraModel):
     id = columns.UUID(primary_key=True, default=uuid4)
@@ -31,6 +23,9 @@ class Answers(DjangoCassandraModel):
     is_commitment = columns.Boolean(required=True)  # Boolean pour savoir si engagement ou pas
     score_response = columns.Double(required=False)  # Score de la question
 
+    @classmethod
+    def get_by_id(cls, idAnswer):
+        return cls.objects.get(pk=idAnswer)
 
 class Challenges(DjangoCassandraModel):
     id = columns.UUID(primary_key=True, default=uuid4)
