+++ backend/questions/views.py
@@ -1,5 +1,4 @@
 from django.http import JsonResponse
-from django.shortcuts import render
 from django.views.decorators.http import require_GET
 
 from questions.models import Challenges, SubChallenges, Questions, Choices
