+++ backend/users/models.py
@@ -9,6 +9,10 @@ class Users(DjangoCassandraModel):
     role = columns.Text(required=True)  # Role as 'employee' or 'client'
     id_client_information = columns.UUID(required=False)  # Optional field for 'client' role
 
+    @classmethod
+    def get_by_id(cls, idUser):
+        return cls.objects.get(pk=idUser)
+
 # Create this before creating User
 class ClientInformation(DjangoCassandraModel):
     id_user = columns.UUID(primary_key=True)
